Quadratic Interpolation Implementation
First, no so good implementation.
TODO:
- Update to tell when number of point is less!
- Binary search

+++ src/main.cpp
@@ -10,7 +10,7 @@ int main()
 
 	try
 	{
-		TaskManager<int> manager("a", "b");
+		TaskManager<float> manager("a", "b");
 		manager.Solve();
 	}
 	catch (const std::exception& ex)
